requirements.txt
fastapi
uvicorn
motor
pydantic
dotenv
httpx
google-genai
pymongo
weasyprint
aiogram
pydantic
dotenv